Remove the RoomLifecycleContributorChannel protocol
As mentioned in 0cb474b.

Resolves #244.
