+++ Sources/AblyChat/DefaultRoomLifecycleContributor.swift
@@ -1,11 +1,11 @@
 import Ably
 
 internal actor DefaultRoomLifecycleContributor: RoomLifecycleContributor, EmitsDiscontinuities, CustomDebugStringConvertible {
-    internal nonisolated let channel: DefaultRoomLifecycleContributorChannel
+    internal nonisolated let channel: any InternalRealtimeChannelProtocol
     internal nonisolated let feature: RoomFeature
     private var discontinuitySubscriptions = SubscriptionStorage<DiscontinuityEvent>()
 
-    internal init(channel: DefaultRoomLifecycleContributorChannel, feature: RoomFeature) {
+    internal init(channel: any InternalRealtimeChannelProtocol, feature: RoomFeature) {
         self.channel = channel
         self.feature = feature
     }
@@ -24,44 +24,5 @@ internal actor DefaultRoomLifecycleContributor: RoomLifecycleContributor, EmitsD
 
     internal nonisolated var debugDescription: String {
         "(\(id): \(feature), \(channel))"
-    }
-}
-
-internal final class DefaultRoomLifecycleContributorChannel: RoomLifecycleContributorChannel, CustomDebugStringConvertible {
-    private let underlyingChannel: any InternalRealtimeChannelProtocol
-
-    internal init(underlyingChannel: any InternalRealtimeChannelProtocol) {
-        self.underlyingChannel = underlyingChannel
-    }
-
-    internal func attach() async throws(InternalError) {
-        try await underlyingChannel.attach()
-    }
-
-    internal func detach() async throws(InternalError) {
-        try await underlyingChannel.detach()
-    }
-
-    internal var state: ARTRealtimeChannelState {
-        underlyingChannel.state
-    }
-
-    internal var errorReason: ARTErrorInfo? {
-        underlyingChannel.errorReason
-    }
-
-    internal func subscribeToState() async -> Subscription<ARTChannelStateChange> {
-        let subscription = Subscription<ARTChannelStateChange>(bufferingPolicy: .unbounded)
-        let eventListener = underlyingChannel.on { subscription.emit($0) }
-        subscription.addTerminationHandler { [weak underlyingChannel] in
-            underlyingChannel?.unsubscribe(eventListener)
-        }
-        return subscription
-    }
-
-    // MARK: - CustomDebugStringConvertible
-
-    internal var debugDescription: String {
-        "\(underlyingChannel)"
     }
 }
